waste-collection-amberg/requirements.txt
pdf2image
pillow
pandas
numpy
easyocr
holidays